requirements.txt
streamlit==1.36.0
pyodbc==5.1.0
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.5